+++ includes/class-map.php
@@ -4,7 +4,7 @@ if (!defined('ABSPATH')) {
     exit;
 }
 
-class IRM_Map
+class ERM_Map
 {
     /**
      * Render the map template.
@@ -14,12 +14,12 @@ class IRM_Map
      */
     public static function render($stops)
     {
-        $marker_color = get_option(IRM_Plugin::OPTION_MARKER_COLOR, IRM_Plugin::DEFAULT_MARKER_COLOR);
-        $route_color  = get_option(IRM_Plugin::OPTION_ROUTE_COLOR, IRM_Plugin::DEFAULT_ROUTE_COLOR);
+        $marker_color = get_option(ERM_Plugin::OPTION_MARKER_COLOR, ERM_Plugin::DEFAULT_MARKER_COLOR);
+        $route_color  = get_option(ERM_Plugin::OPTION_ROUTE_COLOR, ERM_Plugin::DEFAULT_ROUTE_COLOR);
 
         ob_start();
 
-        include IRM_PLUGIN_PATH . 'templates/map.php';
+        include ERM_PLUGIN_PATH . 'templates/map.php';
 
         return ob_get_clean();
     }