+++ includes/class-plugin.php
@@ -4,10 +4,10 @@ if (!defined('ABSPATH')) {
     exit;
 }
 
-class IRM_Plugin
+class ERM_Plugin
 {
-    const OPTION_MARKER_COLOR = 'irm_marker_color';
-    const OPTION_ROUTE_COLOR = 'irm_route_color';
+    const OPTION_MARKER_COLOR = 'erm_marker_color';
+    const OPTION_ROUTE_COLOR = 'erm_route_color';
     const DEFAULT_MARKER_COLOR = '#3d874d';
     const DEFAULT_ROUTE_COLOR = '#3388ff';
 
@@ -24,13 +24,13 @@ class IRM_Plugin
 
     private function __construct()
     {
-        require_once IRM_PLUGIN_PATH . 'includes/class-admin.php';
-        require_once IRM_PLUGIN_PATH . 'includes/class-acf.php';
-        require_once IRM_PLUGIN_PATH . 'includes/class-map.php';
-        require_once IRM_PLUGIN_PATH . 'includes/class-shortcode.php';
+        require_once ERM_PLUGIN_PATH . 'includes/class-admin.php';
+        require_once ERM_PLUGIN_PATH . 'includes/class-acf.php';
+        require_once ERM_PLUGIN_PATH . 'includes/class-map.php';
+        require_once ERM_PLUGIN_PATH . 'includes/class-shortcode.php';
 
-        new IRM_Admin();
-        new IRM_Shortcode();
+        new ERM_Admin();
+        new ERM_Shortcode();
 
         add_action('wp_enqueue_scripts', [$this, 'enqueue_frontend']);
         add_action('admin_enqueue_scripts', [$this, 'enqueue_admin']);
@@ -39,21 +39,21 @@ class IRM_Plugin
     public function enqueue_frontend()
     {
         wp_enqueue_style(
-            'irm-leaflet',
+            'erm-leaflet',
             'https://unpkg.com/leaflet@1.9.4/dist/leaflet.css',
             [],
             '1.9.4'
         );
 
         wp_enqueue_style(
-            'irm-style',
-            IRM_PLUGIN_URL . 'assets/css/easy-route-map.css',
-            ['irm-leaflet'],
-            filemtime(IRM_PLUGIN_PATH . 'assets/css/easy-route-map.css')
+            'erm-style',
+            ERM_PLUGIN_URL . 'assets/css/easy-route-map.css',
+            ['erm-leaflet'],
+            filemtime(ERM_PLUGIN_PATH . 'assets/css/easy-route-map.css')
         );
 
         wp_enqueue_script(
-            'irm-leaflet',
+            'erm-leaflet',
             'https://unpkg.com/leaflet@1.9.4/dist/leaflet.js',
             [],
             '1.9.4',
@@ -61,10 +61,10 @@ class IRM_Plugin
         );
 
         wp_enqueue_script(
-            'irm-script',
-            IRM_PLUGIN_URL . 'assets/js/easy-route-map.js',
-            ['irm-leaflet'],
-            filemtime(IRM_PLUGIN_PATH . 'assets/js/easy-route-map.js'),
+            'erm-script',
+            ERM_PLUGIN_URL . 'assets/js/easy-route-map.js',
+            ['erm-leaflet'],
+            filemtime(ERM_PLUGIN_PATH . 'assets/js/easy-route-map.js'),
             true
         );
     }
@@ -76,10 +76,10 @@ class IRM_Plugin
         }
 
         wp_enqueue_style(
-            'irm-admin',
-            IRM_PLUGIN_URL . 'assets/css/admin.css',
+            'erm-admin',
+            ERM_PLUGIN_URL . 'assets/css/admin.css',
             [],
-            filemtime(IRM_PLUGIN_PATH . 'assets/css/admin.css')
+            filemtime(ERM_PLUGIN_PATH . 'assets/css/admin.css')
         );
     }
 }